+++ galsim/bandpass.py
@@ -399,7 +399,7 @@ class Bandpass(object):
         return Bandpass(self._orig_tp, self.wave_type, self.blue_limit, self.red_limit,
                         new_zeropoint, self.wave_list, self._tp)
 
-    def truncate(self, blue_limit=None, red_limit=None, relative_throughput=None):
+    def truncate(self, blue_limit=None, red_limit=None, relative_throughput=None, preserve_zp='auto'):
         """Return a bandpass with its wavelength range truncated.
 
         This function truncate the range of the bandpass either explicitly (with `blue_limit` or
@@ -413,14 +413,37 @@ class Bandpass(object):
         This function does not remove any intermediate wavelength ranges, but see thin() for
         a method that can thin out the intermediate values.
 
-        @param blue_limit       Truncate blue side of bandpass here. [default: None]
-        @param red_limit        Truncate red side of bandpass here. [default: None]
+        When truncating a bandpass that already has an assigned zeropoint, there are several
+        possibilities for what should happen to the new (returned) bandpass by default.  If red
+        and/or blue limits are given, then the new bandpass will have no assigned zeropoint because
+        it is difficult to predict what should happen if the bandpass is being arbitrarily
+        truncated.  If `relative_throughput` is given, often corresponding to low-level truncation
+        that results in little change in observed quantities, then the new bandpass is assigned the
+        same zeropoint as the original.  This default behavior is called 'auto'.  The user can also
+        give boolean True or False values.
+
+        @param blue_limit           Truncate blue side of bandpass here. [default: None]
+        @param red_limit            Truncate red side of bandpass here. [default: None]
         @param relative_throughput  Truncate leading or trailing wavelengths that are below
-                                this relative throughput level.  (See above for details.)
-                                [default: None]
+                                    this relative throughput level.  (See above for details.)
+                                    [default: None]
+        @param preserve_zp          If True, the new truncated Bandpass will be assigned the same
+                                    zeropoint as the original.  If False, the new truncated Bandpass
+                                    will have a zeropoint of None. If 'auto', the new truncated
+                                    Bandpass will have the same zeropoint as the original when
+                                    truncating using `relative_throughput`, but will have a
+                                    zeropoint of None when truncating using 'blue_limit' and/or
+                                   'red_limit'.  [default: 'auto']
 
         @returns the truncated Bandpass.
         """
+        if preserve_zp == 'auto':
+            if relative_throughput is not None: preserve_zp = True
+            else: preserve_zp = False
+        # Check for weird input
+        if preserve_zp is not True and preserve_zp is not False:
+            raise ValueError("Unrecognized input for preserve_zp: %s"%preserve_zp)
+
         if blue_limit is None:
             blue_limit = self.blue_limit
         if red_limit is None:
@@ -440,8 +463,12 @@ class Bandpass(object):
                 "Can only truncate with relative_throughput argument if throughput is "
                 + "a LookupTable")
 
-        return Bandpass(self._orig_tp, self.wave_type, blue_limit, red_limit,
-                        _wave_list=wave_list, _tp=self._tp)
+        if preserve_zp:
+            return Bandpass(self._orig_tp, self.wave_type, blue_limit, red_limit,
+                            _wave_list=wave_list, _tp=self._tp, zeropoint=self.zeropoint)
+        else:
+            return Bandpass(self._orig_tp, self.wave_type, blue_limit, red_limit,
+                            _wave_list=wave_list, _tp=self._tp)
 
     def thin(self, rel_err=1.e-4, trim_zeros=True, preserve_range=True, fast_search=True,
              preserve_zp=True):