+++ CHANGELOG.md
@@ -107,7 +107,8 @@ New Features
 - Added the FourierSqrt operator to compute the Fourier-space square root of a
   profile.  This is useful in implementing optimal coaddition algorithms; see
   make_coadd.py in the examples directory (#748).
-- Made Bandpass.thin() preserve the zeropoint by default. (#711)
+- Made Bandpass.thin() and Bandpass.truncate() preserve the zeropoint by 
+  default. (#711)
 - Added version information to the compiled C++ library. (#750)